requirements.txt
imageio
numpy
pillow
